Database Update
Updated to new structure as shown on the whiteboard and updated all the data to match. I'm working on a way to list all activity types and get activities by their type. Activity types should be easy to list through the $types array, just get the $types['typeName'] values in a foreach loop and it should work.